Added the single-file header-only idx2.hpp

+++ src/idx2_error.h
@@ -17,7 +17,7 @@ struct error {
   error(t CodeIn, bool StrGenedIn = false, cstr MsgIn = "");
   idx2_T(u) error(const error<u>& Err);
   inline thread_local static cstr Files[64]; // Store file names up the stack
-  inline thread_local static i16 Lines[64]; // Store line numbers up the stack
+  inline thread_local static int Lines[64]; // Store line numbers up the stack
   operator bool() const;
 }; // struct err_template
